+++ InfinityHammer/behavior/Selection.cs
@@ -104,7 +104,7 @@ public class PieceTableSetSelected
     var index = __instance.GetSelectedIndex();
     var piece = __instance.GetPiece((int)__instance.m_selectedCategory, index);
     if (piece && piece.GetComponent<ZNetView>())
-      Selection.CreateGhost(new ObjectSelection(Utils.GetPrefabName(piece.gameObject), false));
+      Selection.CreateGhost(new ObjectSelection(piece.GetComponent<ZNetView>(), false));
   }
 }
 ///<summary>Overrides the piece selection.</summary>